+++ netbox/circuits/filters.py
@@ -3,8 +3,9 @@ from django.db.models import Q
 
 from dcim.filters import CableTerminationFilterSet
 from dcim.models import Region, Site, SiteGroup
+from extras.filters import TagFilter
 from tenancy.filters import TenancyFilterSet
-from utilities.filters import TagFilter, TreeNodeMultipleChoiceFilter
+from utilities.filters import TreeNodeMultipleChoiceFilter
 from utilities.filtersets import ChangeLoggedModelFilterSet, OrganizationalModelFilterSet, PrimaryModelFilterSet
 from .choices import *
 from .models import *
